app: rename msi_receiver to msi_message
... because the MSI receiver code is gone.
